no ability to delete pantry, added style to expiration date

+++ React_Native_Inventory_Tracker/screens/Sections.js
@@ -1,4 +1,4 @@
-import React, {useCallback, useEffect, useState} from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 import { StyleSheet, Text, View, Image, SafeAreaView, Button, SectionList, FlatList, TouchableOpacity, TouchableHighlight, TextInput, Switch, ImageBackground, Alert, Platform, TouchableWithoutFeedback, ScrollView } from "react-native";
 import FloatingButton from '../FloatingButton';
 import * as SQLite from 'expo-sqlite';
@@ -6,25 +6,26 @@ import * as SQLite from 'expo-sqlite';
 
 const db = SQLite.openDatabase('db');
 
-function removeSection(sectionID){
-  db.transaction((tx) => {
-    tx.executeSql(
-      'delete from Section where sectionID = ?;',
-      [sectionID],
-      (tx, results) => {
-        if (results.rowsAffected>0){
-          console.log(results)
-          return(
-            Alert.alert(
-              "",
-              "Section has been deleted",
-              [{
-                text: "Ok",
-                onPress: console.log("Success!")
-              }]
+function removeSection(sectionID) {
+  if (sectionID != 0) {
+    db.transaction((tx) => {
+      tx.executeSql(
+        'delete from Section where sectionID = ?;',
+        [sectionID],
+        (tx, results) => {
+          if (results.rowsAffected > 0) {
+            console.log(results)
+            return (
+              Alert.alert(
+                "",
+                "Section has been deleted",
+                [{
+                  text: "Ok",
+                  onPress: console.log("Success!")
+                }]
+              )
             )
-          )
-        } else{
+          } else {
             return (
               Alert.alert(
                 "",
@@ -35,11 +36,19 @@ function removeSection(sectionID){
                 }]
               )
             )
+          }
         }
-      } 
-    )
-  }); 
-  
+      )
+    });
+  } else {
+    return (Alert.alert(
+      "",
+      "Your pantry can't be deleted...Try to delete any other section!",
+      [{
+        text: "Ok",
+      }]
+    ));
+  }
 }
 
 function getSection() {
@@ -61,97 +70,97 @@ function getSection() {
 }
 
 function Sections({ navigation }) {
-    const sections = getSection()
-  
-    return (
-      <ImageBackground
-        source={require('../assets/cart.jpg')}
-        style={{ width: '100%', height: '100%' }}
-      >
-        <View>
-          <Text style={styles.textHead}>Delete Section: </Text>
-          <FlatList
-            data={sections}
-            keyExtractor={(item, index) => index}
-            renderItem={({ item, index, separators }) =>
-              <TouchableOpacity style={styles.sections} 
-                onPress={() =>
-                  Alert.alert(
-                    "Are you sure you want to delete this section?",
-                    "You cannot undo this action.",
-                    [
-                      {
-                        text: "No",
-                      },
-                      {
-                        text: "Yes",
-                        onPress: () => removeSection(item.sectionID)
-                      }])}>
-                <Text style={styles.text}>{item.sectionName}</Text>
-              </TouchableOpacity>
-            }
-          />
-          <TouchableOpacity
-            style={styles.addToWishList}
-            onPress={() => navigation.navigate('AddSection')}>
-            <Text style={styles.text} >Add Section</Text>
-          </TouchableOpacity>
-        </View>
-        <FloatingButton
-          style={styles.floatinBtn}
-          onPress={() => navigation.navigate('INVENTORY TRACKING APP')}
+  const sections = getSection()
+
+  return (
+    <ImageBackground
+      source={require('../assets/cart.jpg')}
+      style={{ width: '100%', height: '100%' }}
+    >
+      <View>
+        <Text style={styles.textHead}>Delete Section: </Text>
+        <FlatList
+          data={sections}
+          keyExtractor={(item, index) => index}
+          renderItem={({ item, index, separators }) =>
+            <TouchableOpacity style={styles.sections}
+              onPress={() =>
+                Alert.alert(
+                  "Are you sure you want to delete this section?",
+                  "You cannot undo this action.",
+                  [
+                    {
+                      text: "No",
+                    },
+                    {
+                      text: "Yes",
+                      onPress: () => removeSection(item.sectionID)
+                    }])}>
+              <Text style={styles.text}>{item.sectionName}</Text>
+            </TouchableOpacity>
+          }
         />
-      </ImageBackground>
-    );
+        <TouchableOpacity
+          style={styles.addToWishList}
+          onPress={() => navigation.navigate('AddSection')}>
+          <Text style={styles.text} >Add Section</Text>
+        </TouchableOpacity>
+      </View>
+      <FloatingButton
+        style={styles.floatinBtn}
+        onPress={() => navigation.navigate('INVENTORY TRACKING APP')}
+      />
+    </ImageBackground>
+  );
 }
 
 const styles = StyleSheet.create({
-    textHead: {
-        textAlign: 'center',
-        fontSize: 30,
-        fontFamily: 'Avenir',
-        fontWeight: 'bold',
-        color: 'black',
-    },
-    sections: {
-        textAlign: "center",
-        borderWidth: 3,
-        borderColor: "darkgrey",
-        fontSize: 40,
-        color: "black",
-        height: 60,
-        width: 250,
-        left: 70,
-    },
-    text: {
-        textAlign: 'center',
-        fontSize: 14,
-        fontFamily: 'Avenir',
-        fontWeight: 'bold',
-        color: 'black',
-    },
-    addToWishList: {
-        backgroundColor: '#859a9b',
-        borderRadius: 20,
-        padding: 10,
-        shadowColor: '#303838',
-        shadowOffset: { width: 0, height: 5 },
-        shadowRadius: 10,
-        shadowOpacity: 0.35,
-        justifyContent: 'flex-end',
-        top: 200,
-        width: "50%",
-        left: 90,
-    },
-    floatinBtn: {
-        color: 'grey',
-        position: 'absolute',
-        bottom: 10,
-        right: 10,
-    },
-    
+  textHead: {
+    textAlign: 'center',
+    fontSize: 30,
+    fontFamily: 'Avenir',
+    fontWeight: 'bold',
+    color: 'black',
+  },
+  sections: {
+    textAlign: "center",
+    borderWidth: 3,
+    borderColor: "darkgrey",
+    fontSize: 40,
+    color: "black",
+    height: 60,
+    width: 250,
+    left: 70,
+  },
+  text: {
+    textAlign: 'center',
+    fontSize: 14,
+    fontFamily: 'Avenir',
+    fontWeight: 'bold',
+    color: 'black',
+  },
+  addToWishList: {
+    backgroundColor: '#859a9b',
+    borderRadius: 20,
+    padding: 10,
+    shadowColor: '#303838',
+    shadowOffset: { width: 0, height: 5 },
+    shadowRadius: 10,
+    shadowOpacity: 0.35,
+    justifyContent: 'flex-end',
+    top: 200,
+    width: "50%",
+    left: 90,
+  },
+  floatinBtn: {
+    color: 'grey',
+    position: 'absolute',
+    bottom: 10,
+    right: 10,
+  },
+
 
 });
 
 export default Sections;
-export {getSection};
+export { getSection };
